backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
xgboost==2.0.3
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1